// src/app/models/Fields.ts
export interface Fields {
  ebike?: number;
  capacity?: number;
  name: string;
  nom_arrondissement_communes: string;
  numbikesavailable?: number;
  is_installed: string;
  is_renting: string;
  mechanical?: number;
  stationcode: string;
  coordonnees_geo?: number[];
  numdocksavailable?: number;
  duedate: string;
  is_returning: string;
}